move locale to service, move i18n json to schema
